# lib/api-zod/src/generated/types/geminiConversationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiConversationInput {
  title: string;
}